refactor(article): rename article components to article-list scope - Rename ArticleDataTable.vue and columns.ts from article/ to article-list/ directory - Update import paths in admin page and tests accordingly - Adjust mocks in tests to reflect new component paths
